Rename disease profile component
