+++ src/public/disease/Profile.js
